Start to adapt shardingsphere and failed...
